Update hover styles and update task history page
