+++ client/src/components/Header/Header.module.css
@@ -65,7 +65,7 @@
   margin: 5px 10px;
 }
 
-.navDropdownItem:hover{
+.navDropdownItem:hover {
   cursor: pointer;
 }
